8 processes

+++ config/puma.rb
@@ -39,7 +39,7 @@
 
 # Processes: ec2 default just deploys to 1 process engine- 
 
-workers Integer(1)
+workers Integer(8)
 threads_count = Integer(1)
 threads threads_count, threads_count
